Convert long single-line functions into canonical multi-line ones
It isn't a hard convention within the community, but general preference
is to use short function notation for very short statements. The rule of
thumb I myself use is if the code breaks the 80 characters wide, I use
normal function syntax instead of short one. It's just a bit easier to
parse, in my humble opinion.