Move protocol implementations under the modules they are defined for
I find this makes it easier to parse the code and move modules around.
Both of which happen a lot during the early days of the project. As the
number of commands and events grow, we would need to consider splitting
this one file into multiple. And it's easier to do when cohesive logic
is already neatly grouped.