Hide process communication details under Account's public API
It is a common-wide practice in Elixir and Erlang to hide the details of
the message passing behind the modules public API.

This implementation hiding has three concrete benefits - two short term
and one long term:
1. In the short term, hiding details of structs and message passing
   routines allows module users to not worry about aliasing internal
   submodules;
2. Another short term benefit is that public API (as opposed to explicit
   `send` calls) is way easier to abstract and format cleanly. It is,
   for example, possible to create a very clean looking pipelines using
   the public API, as would be demonstrated with the next commit;
3. In the long term, hiding details allows smoother evolution of both
   the process implementation and the message protocol itself.

It is the third benefit that is the main driving force behind this
commit. I am planning to transform Account into a GenServer in couple
of commits. Without public API, we would have to change both
`acount_run.exs` and `account_test.exs` excessively. With public API
both userland modules can stay unchanged.