Convert Account to a GenServer
There is technically nothing wrong with the previous implementation,
outside of it missing logic for tracing, error reporting and interface
to slot it in into supervision trees (`child_spec`).

Commonly in Elixir we do not use `send` and `receive` directly and
instead rely on the OTP and Elixir abstractions such as GenServer, Task
or Agent. Aside of it just being a neat convention, these abstractions
actually come with out-of-the-box interfaces and conveniences for error
handling and supervision.

If I was teaching people Elixir from scratch, I'd show send and receive
as these are fundamental building blocks behind OTP. For canonical use
of Elixir, however, I'd stick with GenServer.