Split commands handling and events application
This is the most opinionated of all the changes and technically has not
much to do with Elixir. But I do like to separate command handling and
event application in event-based processes. This makes it easier to
enforce boundaries and build a reconstitution algorithm when persistence
is added.

I also flipped the arguments around - command/event come first, state
second. This is due to Elixir's standard library working better with
such argument arrangement. For example, `Enum.reduce/3` would always
pass accumulator (state or command) as the second argument to the
applicator function.